move irqs to hal_interrupt, move timer 1 callback to device_callback, move all band heater logic to extrusion task
