+++ Core/layer_1/band_heater.h
@@ -55,7 +55,6 @@ class device::band_heater
         uint32_t initialize(uint8_t arg_temperature_zone_id, uint8_t arg_output_pulse_timer_id, osMutexId_t arg_mutex_handle);
         uint32_t initialize_output_pulse_timer(uint8_t arg_output_pulse_timer_id);
         uint32_t set_period(uint16_t arg_period);
-        uint32_t update_period();
         hal::timer_handle_t* get_output_pulse_timer_module();
         static void zero_crossing_pulse_restart();
         friend void output_pulse_restart(band_heater* arg_band_heater);
