+++ Core/layer_1/device_callback.h
@@ -1,0 +1,37 @@
+/***********************************************************************************************************************
+ * Main_Controller
+ * device_callback.h
+ *
+ * wilson
+ * 10/31/24
+ * 11:19 PM
+ *
+ * Description:
+ *
+ **********************************************************************************************************************/
+
+#ifndef MAIN_CONTROLLER_DEVICE_CALLBACK_H
+#define MAIN_CONTROLLER_DEVICE_CALLBACK_H
+
+/* c/c++ includes */
+
+/* stm32 includes */
+
+/* third-party includes */
+
+/* layer_0_hal includes */
+
+/* layer_1_rtosal includes */
+
+/* layer_2_device includes */
+#include "device.h"
+/* layer_3_control includes */
+
+/* layer_4_sys_op includes */
+
+/* layer_n_meta_structure includes */
+
+void device_callback_tim_1_input_capture_pulse_detected_callback(TIM_HandleTypeDef *htim);
+
+
+#endif //MAIN_CONTROLLER_DEVICE_CALLBACK_H
